Don't expect user to be psychic
If message is split into multiple sentences or user says multiple before
bot processes the messages, bot could ask question in response to one and
then interperate a later message from user as the response even if it was
said before the expectation message was said.

Now bot does only treats message as a response to expectation if newer
than the latest message number at time of processing.